+++ ublox-short-range/src/command/data_mode/responses.rs
@@ -7,14 +7,14 @@ use heapless::{consts, String};
 #[derive(Clone, AtatResp)]
 pub struct ConnectPeerResponse {
     #[at_arg(position = 0)]
-    pub peer_handle: u32,
+    pub peer_handle: usize,
 }
 
 /// 5.5 Peer list +UDLP
 #[derive(Clone, AtatResp)]
 pub struct PeerListResponse {
     #[at_arg(position = 0)]
-    pub peer_handle: u32,
+    pub peer_handle: usize,
     #[at_arg(position = 1)]
     pub protocol: String<consts::U64>,
     #[at_arg(position = 2)]
@@ -27,7 +27,7 @@ pub struct PeerListResponse {
 #[derive(Clone, AtatResp)]
 pub struct BindResponse {
     #[at_arg(position = 0)]
-    pub channel_id_1: u32,
+    pub channel_id_1: usize,
     #[at_arg(position = 1)]
-    pub channel_id_2: u32,
+    pub channel_id_2: usize,
 }
